feat: eliminate all cast_sign_loss warnings (25→0)
Add cast_sign_loss lint to [lints.clippy] and fix all 25 warnings:

RPM parser: Change EntryInfo.tag, InstalledRpmPackage fields
(epoch/size/dir_indexes) from i32 to u32; compute_data_length and
string_tag_length return Option<usize> instead of isize with -1 sentinel;
TYPE_SIZES uses Option<u32> with None for variable-length types; add
positive_offset helper for safe i32→u32 offset conversions.

Scanner/CLI: Replace max_in_memory: i64 with MemoryMode enum
(CollectFirst/StreamUnlimited/Limit(usize)); replace processes: i32 with
ProcessMode enum (Sequential/Auto/Parallel(usize)); change format_size
to accept f64 directly, eliminating f64→u64→f64 round-trip.

License detection: Refactor known_pos (i32 with -1 sentinel) and
pos (i64 with -1 sentinel) to Option<usize> with None; change
unknowns_by_pos/stopwords_by_pos to HashMap<Option<usize>, usize>;
refactor matchable_tokens() from Vec<i32> with -1 sentinels to
Vec<Option<TokenId>>; add HIGH_RESEMBLANCE_THRESHOLD_TENTHS const u32
as canonical source, derive f32 from it; replace u8::try_from().is_ok()
+ as cast with .ok(); add #[allow] for f64→u8 in loader.rs (no clean
alternative); use unsigned range literals in test code.

+++ src/license_detection/models/license_match.rs
@@ -454,7 +454,7 @@ impl LicenseMatch {
         let unknowns_in_match: usize = span
             .iter()
             .filter(|&pos| pos != max_pos)
-            .filter_map(|pos| query.unknowns_by_pos.get(&Some(pos as i32)))
+            .filter_map(|pos| query.unknowns_by_pos.get(&Some(pos)))
             .sum();
         qregion_len + unknowns_in_match
     }
